H264_Encoding: Prepare some data structures for adding packed raw data
Under some encoding scenario, the user-space application hopes that the driver
can insert the passed packed rawdata into the coded clip. But the insertion of
packed rawdata is related with the slice. So some data structures are added so
that it can store how the packed rawdata is inserted into the coded clip
per-slice.

Signed-off-by: Zhao, Yakui <yakui.zhao@intel.com>