Use the inline CPUID assembly to obtain CPUID string instead of /proc/cpuinfo
On some systems there is no access to /proc/cpuinfo. So the inline assembly
is used directly to detect the CPUID string.

V2->V3: Follow the suggestion from Google engineer to remove the unused code
of "if 0". And the "unsigned int" data type is updated to "uint32_t".

V1->V2: Based on the Matt Turner's suggestion the __cpuid defined in GCC
cpuid.h is called directly, which is helpful to handle the PIC issue on
32-bit.

Signed-off-by: Zhao Yakui <yakui.zhao@intel.com>